Updated gpt_oss test file with KV handoff check
