+++ tests/transformers/disaggregated/test_gemma4_moe_disagg_kv_share_w_hf_fp32.py
@@ -22,7 +22,7 @@ from QEfficient import QEFFAutoModelForImageTextToText
 from QEfficient.base.onnx_transforms import FP16ClipTransform
 from QEfficient.generation.cloud_infer import QAICInferenceSession
 
-pytestmark = pytest.mark.skip(reason="Temporarily disabled for PR 1150.")
+pytestmark = pytest.mark.skip(reason="")
 
 MODEL_NAME = "google/gemma-4-26B-A4B-it"
 SYSTEM_PROMPT = "You are a helpful assistant."
