+++ tests/transformers/disaggregated/test_qwen3_5_disagg_kv_share_w_hf_fp32.py
@@ -23,7 +23,7 @@ from transformers import AutoConfig, AutoModelForCausalLM, AutoModelForImageText
 from QEfficient import QEFFAutoModelForImageTextToText
 from QEfficient.generation.cloud_infer import QAICInferenceSession
 
-pytestmark = pytest.mark.skip(reason="Temporarily disabled for PR 1150.")
+pytestmark = pytest.mark.skip(reason="")
 
 MODEL_NAME = "Qwen/Qwen3.6-35B-A3B"
 # MODEL_NAME = 'tiny-random/qwen3.5-moe'
